// old/agir-work-planning-web/src/app/shared/models/findOptions/projectFields.ts
export enum PROJECT_FIELDS {
  INTERVENTION_IDS = 'interventionIds',
  INTERVENTIONS = 'interventions',
  ANNUALDISTRIBUTION_ANNUALPERIODS_PROGRAMBOOKID = 'annualDistribution.annualPeriods.programBookId',
  START_YEAR = 'startYear',
  END_YEAR = 'endYear',
  PROJECT_NAME = 'projectName',
  PROJECT_TYPE_ID = 'projectTypeId',
  ANNUALDISTRIBUTION_ANNUALPERIODS_CATEGORYID = 'annualDistannualDistribution.annualPeriods.categoryId',
  ANNUALDISTRIBUTION_ANNUALPERIODS_STATUS = 'annualDistribution.annualPeriods.status',
  ANNUALDISTRIBUTION_ANNUALPERIODS_ANNUALBUDGET = 'annualDistribution.annualPeriods.annualBudget',
  ANNUALDISTRIBUTION_ANNUALPERIODS_YEAR = 'annualDistribution.annualPeriods.year',
  ANNUALDISTRIBUTION_ANNUALPERIODS_PROGRAMBOOK_ID = 'annualDistribution.annualPeriods.programBook._id',
  ANNUALDISTRIBUTION_ANNUALPERIODS_PROGRAMBOOK_ANNUALPROGRAMID = 'annualDistribution.annualPeriods.programBook.annualProgramId',
  ANNUALDISTRIBUTION_ANNUALPERIODS_PROGRAMBOOK_NAME = 'annualPeriods.programannualDistribution.annualPeriods.programBook.name',
  STATUS = 'status',
  SUBMISSION_NUMBER = 'submissionNumber',
  DRM_NUMBER = 'drmNumber',
  GLOBAL_BUDGET = 'globalBudget',
  EXECUTOR_ID = 'executorId',
  STREET_NAME = 'streetName',
  STREET_FROM = 'streetFrom',
  STREET_TO = 'streetTo',
  BOROUGH_ID = 'boroughId',
  MEDAL_ID = 'medalId',
  DECESION_TYPE_ID = 'decisions.typeId',
  DECESIONS = 'decisions',
  GEOMETRYPIN = 'geometryPin',
  ANNUAL_DISTRIBUTION = 'annualDistribution',
  SUB_CATEGORY_IDS = 'subCategoryIds',
  GLOBAL_BUDGET_ALLOWANCE = 'globalBudget.allowance',
  LENGTH_VALUE = 'length.value',
  ANNULAL_DISTRIBUTION = 'annualDistribution',
  RISK_ID = 'riskId',
  ROAD_NETWORK_TYPE_ID = 'roadNetworkTypeId',
  SERVICE_PRIORITIES = 'servicePriorities',
  IN_CHARGE_ID = 'inChargeId'
}